moving autoprune resource to new file
